añadido autorizacion a usuario y adiestrador, mejorado eliminacion de…

+++ controllers/user.controller.js
@@ -4,14 +4,28 @@ const jwt = require('jsonwebtoken');
 
 // import internal modules
 const userService = require('../services/user.service');
-const { JWT_PASSPHRASE } = require('../util/auth.config');
+const { JWT_PASSPHRASE, AUTHORITIES } = require('../util/auth.config');
 
 exports.findAll = async (req, res, next) => {
-  const users = await userService.findAll();
-  res.status(200).send(users);
+  if (req.requesterData.role === AUTHORITIES.GOD) {
+    const users = await userService.findAll();
+    res.status(200).send(users);
+  } else {
+    const error = new Error('Unauthorized');
+    error.httpStatus = 403;
+    next(error);
+  }
 };
 
 exports.findById = async (req, res, next) => {
+  if (
+    req.requesterData.userId !== req.params.userId &&
+    req.requesterData.role !== AUTHORITIES.GOD
+  ) {
+    const error = new Error('Unauthorized');
+    error.httpStatus = 403;
+    next(error);
+  }
   try {
     const user = await userService.findById(req.params.userId);
     res.status(200).send(user);
@@ -32,15 +46,29 @@ exports.create = async (req, res, next) => {
 };
 
 exports.deleteById = async (req, res, next) => {
+  if (req.requesterData.role !== AUTHORITIES.GOD) {
+    const error = new Error('Unauthorized');
+    error.httpStatus = 403;
+    next(error);
+  }
   try {
-    await userService.deleteById(req.params.userId);
-    res.status(204).send();
+    const result = await userService.deleteById(req.params.userId);
+    if (result) res.status(204).send();
+    else res.status(404).send();
   } catch (err) {
     next(err);
   }
 };
 
 exports.update = async (req, res, next) => {
+  if (
+    req.requesterData.userId !== req.params.userId &&
+    req.requesterData.role !== AUTHORITIES.GOD
+  ) {
+    const error = new Error('Unauthorized');
+    error.httpStatus = 403;
+    next(error);
+  }
   try {
     const userActualizado = await userService.update(
       req.params.userId,
@@ -60,7 +88,8 @@ exports.generateToken = async (req, res, next) => {
       const token = jwt.sign(
         {
           userId: user._id,
-          email: user.email
+          email: user.email,
+          role: user.role
         },
         JWT_PASSPHRASE,
         { expiresIn: '1h' }
